Split DataStream.cs into per-family files
Move text-stream types into Runtime/Scripts/TextDataStream.cs and
byte-stream types into Runtime/Scripts/ByteDataStream.cs. DataStream.cs
now holds only the shared StreamInfo, StreamError,
ReadIncrementalInstructionBase<T>, StreamOptions, and
StreamHandlerRegistry.

Pure relocation — no behaviour, API, or namespace changes. The existing
DataStream.cs.meta GUID is preserved; the two new files get fresh
GUIDs.
