+++ Runtime/Scripts/ByteDataStream.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3031e9f2add84ec79bc2c3c3aa4dc594
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
