+++ Runtime/Scripts/TextDataStream.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 404b9d1fe0b74071a087d73c95700f92
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
